+++ tests/fixtures.py
@@ -14,6 +14,13 @@ PATH_TO_GENERIC_0_3_1_FILE = os.path.join(
     "MA20123456__2020_08_17_145752__B3.h5",
 )
 
+PATH_TO_GENERIC_0_4_1_FILE = os.path.join(
+    PATH_OF_CURRENT_FILE,
+    "h5",
+    "v0.4.1",
+    "MA190190000__2021_01_19_011931__C3.h5",
+)
+
 
 @pytest.fixture(scope="function", name="generic_well_file")
 def fixture_generic_well_file():