+++ tests/test_file_migration.py
@@ -8,13 +8,17 @@ from mantarray_file_manager import BACKEND_LOG_UUID
 from mantarray_file_manager import BARCODE_IS_FROM_SCANNER_UUID
 from mantarray_file_manager import COMPUTER_NAME_HASH_UUID
 from mantarray_file_manager import CURRENT_HDF5_FILE_FORMAT_VERSION
+from mantarray_file_manager import FILE_VERSION_PRIOR_TO_MIGRATION_UUID
 from mantarray_file_manager import file_writer
 from mantarray_file_manager import IS_FILE_ORIGINAL_UNTRIMMED_UUID
 from mantarray_file_manager import MantarrayH5FileCreator
 from mantarray_file_manager import migrate_to_next_version
+from mantarray_file_manager import ORIGINAL_FILE_VERSION_UUID
 from mantarray_file_manager import TRIMMED_TIME_FROM_ORIGINAL_END_UUID
 from mantarray_file_manager import TRIMMED_TIME_FROM_ORIGINAL_START_UUID
 from mantarray_file_manager import UnsupportedFileMigrationPath
+from mantarray_file_manager import UTC_TIMESTAMP_OF_FILE_VERSION_MIGRATION_UUID
+from mantarray_file_manager import WELL_FILE_CLASSES
 from mantarray_file_manager import WELL_INDEX_UUID
 from mantarray_file_manager import WELL_NAME_UUID
 from mantarray_file_manager import WellFile_0_3_1
@@ -24,6 +28,7 @@ import pytest
 from stdlib_utils import get_current_file_abs_directory
 
 from .fixtures import PATH_TO_GENERIC_0_3_1_FILE
+from .fixtures import PATH_TO_GENERIC_0_4_1_FILE
 
 PATH_OF_CURRENT_FILE = get_current_file_abs_directory()
 
@@ -56,7 +61,6 @@ def test_migrate_to_next_version__When_invoked_on_a_0_3_1_file_with_no_working_d
         assert new_file_path.startswith(tmp_dir)
 
 
-@freeze_time("2021-01-18 13:45:30.543221")
 def test_migrate_to_next_version__When_invoked_on_a_0_3_1_file__Then_the_new_file_has_features_of_0_4_1__including_raw_data_old_metadata_and_new_metadata():
     with tempfile.TemporaryDirectory() as tmp_dir:
         new_file_path = migrate_to_next_version(
@@ -79,6 +83,53 @@ def test_migrate_to_next_version__When_invoked_on_a_0_3_1_file__Then_the_new_fil
         assert wf.get_h5_attribute(str(TRIMMED_TIME_FROM_ORIGINAL_END_UUID)) == 0
         assert wf.get_h5_attribute(str(BACKEND_LOG_UUID)) == ""
         assert wf.get_h5_attribute(str(COMPUTER_NAME_HASH_UUID)) == ""
+
+        # raw data
+        np.testing.assert_array_equal(
+            wf.get_raw_tissue_reading(), old_wf.get_raw_tissue_reading()
+        )
+        np.testing.assert_array_equal(
+            wf.get_raw_reference_reading(), old_wf.get_raw_reference_reading()
+        )
+
+        wf.get_h5_file().close()  # safe clean-up when running CI on windows systems
+        old_wf.get_h5_file().close()  # safe clean-up when running CI on windows systems
+
+
+@freeze_time("2021-01-18 13:45:30.543221")
+def test_migrate_to_next_version__When_invoked_on_a_0_4_1_file__Then_the_new_file_has_features_of_0_4_2__including_raw_data_old_metadata_and_new_metadata():
+    old_version = "0.4.1"
+    new_version = "0.4.2"
+
+    with tempfile.TemporaryDirectory() as tmp_dir:
+        new_file_path = migrate_to_next_version(
+            PATH_TO_GENERIC_0_4_1_FILE, working_directory=tmp_dir
+        )
+
+        wf = WELL_FILE_CLASSES[new_version](new_file_path)
+        old_wf = WELL_FILE_CLASSES[old_version](PATH_TO_GENERIC_0_4_1_FILE)
+        assert wf.get_file_version() == new_version
+
+        # old metadata (since it is all copied by default, testing a subset seems reasonable for now)
+        assert (
+            wf.get_h5_attribute(str(COMPUTER_NAME_HASH_UUID))
+            == "8a1f943c08bf5de71ab50509532a6869962ce2b740d7491f8baf38a9caf12893df4507be47c69c74b90e9a1496d5952e891650561817e78e9388f2f63eb28065"
+        )
+        assert (
+            wf.get_h5_attribute(str(BACKEND_LOG_UUID))
+            == "a7f0199e-362a-41ca-9d4f-6c85f7df9cc7"
+        )
+
+        # new metadata
+        assert wf.get_h5_attribute(str(ORIGINAL_FILE_VERSION_UUID)) == ""
+        assert (
+            wf.get_h5_attribute(str(FILE_VERSION_PRIOR_TO_MIGRATION_UUID))
+            == old_version
+        )
+        assert (
+            wf.get_h5_attribute(str(UTC_TIMESTAMP_OF_FILE_VERSION_MIGRATION_UUID))
+            == "2021-01-18 13:45:30.543221"
+        )
 
         # raw data
         np.testing.assert_array_equal(